requirements.txt
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
gunicorn
whitenoise
psycopg2-binary
